data_generation/requirements.txt
openml
pytorch
scikit-learn
pytorch-lightning
py-experimenter